pull out group test into its own module
